fix: harden libpod client and split container types module
- client.rs: replace .expect() on URI parse with proper Result propagation
- client.rs: replace .unwrap() in Windows pipe retry with safe fallback
- engine/health.rs: URL-encode container names in inspect paths
- engine/mod.rs: URL-encode container name in exec path
- libpod/types/container: split 571-line file into spec.rs + response.rs
  under container/ (hard 500-line limit); re-export all types via mod.rs
