+++ internal/engine/mod.rs
@@ -82,7 +82,7 @@ impl Engine {
 			..Default::default()
 		};
 
-		let path = format!("/libpod/containers/{container_name}/exec");
+		let path = format!("/libpod/containers/{}/exec", crate::libpod::urlencoded(container_name));
 		let resp: crate::libpod::types::exec::ExecCreateResponse =
 			self.client.post_json(&path, &exec_cfg).await.map_err(ComposeError::Podman)?;
 		let exec_id = resp.id;
